[Java]: Use long for offset to enable the use of large log buffers in FileBlockHandler.

+++ aeron-client/src/main/java/uk/co/real_logic/aeron/Connection.java
@@ -181,7 +181,7 @@ public class Connection
         final int bytesConsumed = resultingOffset - termOffset;
         if (resultingOffset > termOffset)
         {
-            final long offset = (capacity * activeIndex) + termOffset;
+            final long offset = ((long)capacity * activeIndex) + termOffset;
             try
             {
                 fileBlockHandler.onBlock(logBuffers.fileChannel(), offset, bytesConsumed, sessionId);
